refactor: modularize components, add ErrorBoundary, and improve error handling
