+++ package.json
@@ -1,6 +1,7 @@
 {
     "name": "tubesize-workspace",
     "private": true,
+    "packageManager": "pnpm@10.30.3+sha512.c961d1e0a2d8e354ecaa5166b822516668b7f44cb5bd95122d590dd81922f606f5473b6d23ec4a5be05e7fcd18e8488d47d978bbe981872f1145d06e9a740017",
     "scripts": {
         "prepare": "husky",
         "prettier:write": "prettier . --write",
