+++ tubesize/src/components/options/headerOptions.tsx
@@ -1,0 +1,12 @@
+import "../../styles/options.css";
+
+export default function HeaderOptions() {
+    return (
+        <div className="header">
+            <a id="backBtn" href="index.html">
+                &larr; Back
+            </a>
+            <h3>Options</h3>
+        </div>
+    );
+}
