+++ tubesize/src/pages/error.tsx
@@ -1,0 +1,24 @@
+import { useState } from "react";
+import Options from "../pages/options";
+import Header from "../components/header";
+
+interface Props {
+    errorMessage: string;
+}
+
+function ErrorPage({ errorMessage }: Props) {
+    const [useOptionsPage, setUseOptionsPage] = useState(false);
+    if (useOptionsPage) return <Options />;
+    return (
+        <>
+            <Header videoData={null} setUseOptionsPage={setUseOptionsPage} />
+            <div className="error-page">
+                <div className="error-page-icon">⚠</div>
+                <div className="error-page-title">Something went wrong</div>
+                <div className="error">{errorMessage}</div>
+            </div>
+        </>
+    );
+}
+
+export default ErrorPage;
