+++ tubesize/src/pages/options.tsx
@@ -1,0 +1,54 @@
+import "../styles/options.css";
+import CONFIG from "../constants";
+import HeaderOptions from "../components/options/headerOptions";
+import OptionItem from "../components/options/optionItem";
+import CacheSettings from "../components/options/cacheSettings";
+import ApiFallbackSetting from "../components/options/apiFallbackSetting";
+import { useState, useEffect } from "react";
+import { getFromSyncCache } from "../cache";
+
+export default function Options() {
+    const [optionsState, setOptionsState] = useState<Record<any, any>>();
+
+    useEffect(() => {
+        (async () => {
+            const options = await getFromSyncCache();
+            setOptionsState(options);
+        })();
+    }, []);
+
+    return (
+        <>
+            <HeaderOptions />
+            <div className="container">
+                <div className="description">Select which resolutions to display:</div>
+                <div id="options-grid">
+                    {CONFIG.optionIDs.map((option) => {
+                        return (
+                            <OptionItem
+                                key={option}
+                                option={option}
+                                optionsState={optionsState}
+                                setOptionsState={setOptionsState}
+                            />
+                        );
+                    })}
+                </div>
+            </div>
+
+            <div className="section-divider"></div>
+            <CacheSettings />
+
+            <div className="section-divider"></div>
+            <ApiFallbackSetting />
+
+            <div className="section-divider"></div>
+            <div className="author">
+                <a href="https://github.com/MohamedSayed0573" target="_blank">
+                    <img src="icons/github.svg" alt="" width="14" height="14" />
+                    @Mohamed Sayed
+                </a>
+            </div>
+        </>
+    );
+}
