+++ tubesize/src/styles/options.css
@@ -24,15 +24,11 @@ body {
     color: #fff;
 }
 #backBtn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
     background: rgba(255, 255, 255, 0.05);
     color: #fff;
     border: 1px solid rgba(255, 255, 255, 0.1);
     border-radius: 4px;
     cursor: pointer;
-    transition: all 0.2s;
     padding: 4px 6px;
     font-size: 12px;
     text-decoration: none;
@@ -246,6 +242,11 @@ input[type="checkbox"]:checked {
         visibility 0.16s ease;
     z-index: 5;
 }
+.apiFallbackSetting {
+    color: inherit;
+    text-decoration: underline;
+}
+
 .info-tooltip:hover .info-tooltip-content,
 .info-tooltip:focus-within .info-tooltip-content {
     opacity: 1;
